Completed upload for submit command
